pyproject.toml
[project]
name = "gradio-mssql-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.5.0",
    "google-genai>=1.47.0",
    "gradio>=5.49.1",
    "ipykernel>=7.1.0",
    "nbclient>=0.10.2",
    "nbformat>=5.10.4",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "soundfile>=0.13.1",
    "sqlalchemy>=2.0.44",
    "statsmodels>=0.14.5",
    "tabulate>=0.9.0",
    "vl-convert-python>=1.8.0",
]
